Fix sooperlooper per-loop bindings
Added ctrl:loop controllers for all per-loop bindable controllers (as
previously defined, there may benefit more).

But also simple 'ctrl' controllers to always handle the selected loop.

The mode of the binding to be made is still changed with
selected_loop_cc_binding (on the last page of controls).

It is indicated on per-loop bindable controls with (loop number) added
to their name.

This is updated whenever the loop selection is changed or when
selected_loop_cc_binding is set to True.

The earlier implementation was incorrect in multiple aspects:

1. It bound ctrl:loopnumbers only, and it only worked when the loop was
also already the selected one.
   This check is now no longer made.
2. Checks for current state checked the selected loop, instead of the
targeted loop. The targeted loop (chan) is now set and checked.
2. Selecting a loop with another method than going to the loop
selection page and back to another controller page did not work
because the UI did not refresh the actual controllers.
   Fixed this by calling
      self.state_manager.send_cuia("refresh_screen", ["control"])
   (I did not find a more direct way)