Fixes first part of zynthian/zynthian-issue-tracking#1317.
Revert (most) symbol names.
Only add array suffix if part of an array.
Only add channel suffix if a multichannel parameter.
Only add "switch" or "enum" suffix if the symbol is not unique without them.
There will be some parameters that were previously broken or bodged which may need user to manually reconfigure, but not many.